Make setup.py not import PyOpenSSL by moving the implementation
